% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_approximation.R
\name{avoid_zero_w_part}
\alias{avoid_zero_w_part}
\title{adjust w_part until it is shifted away from zero}
\usage{
avoid_zero_w_part(x, y, lambdas, weights, s0, increment = 0.01)
}
\arguments{
\item{x}{tested summed score}

\item{y}{observed summed score}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}

\item{s0}{initial s value}

\item{increment}{value to increment y value per iteration}
}
\value{
list of s, mu and w_part values
}
\description{
adjust w_part until it is shifted away from zero
}

