% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulant_generating_functions.R
\name{cgf_0}
\alias{cgf_0}
\title{cumulant generating function, see appendix A of Jiang et al 2015 AJHG 97:272}
\usage{
cgf_0(mu, s, lambdas, weights)
}
\arguments{
\item{mu}{mutation rate}

\item{s}{something}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}
}
\value{
summed value
}
\description{
This estimates the joint CGF of (Y,Z) as:
}
\details{
\eqn{K_{Y,Z}(t,s) = \sum_{l=1}^{p} \lambda_l (e^{c_l t + s} - 1)}
}

