% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulant_generating_functions.R
\name{cgf_2}
\alias{cgf_2}
\title{cumulant generating function, second derivative}
\usage{
cgf_2(mu, s, lambdas, weights)
}
\arguments{
\item{mu}{mutation rate}

\item{s}{something}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}
}
\value{
determinant of the second derivative
}
\description{
The second derivative is calculated as:
}
\details{
\eqn{K^{''}(t,s) = \begin{bmatrix}
     \sum_{l=1}^{p} \lambda_l {c_l}^2 (e^{c_l t + s} - 1) &
            \sum_{l=1}^{p} \lambda_l {c_l} (e^{c_l t + s} - 1) \\
     \sum_{l=1}^{p} \lambda_l {c_l} (e^{c_l t + s} - 1) &
            \sum_{l=1}^{p} \lambda_l (e^{c_l t + s} - 1) \\
   \end{bmatrix}}

NOTE: The values calculated by this function differ very slightly from the
      original package function, due to floating point precision, given how
      the calculation is performed. This shouldn't change the overall result,
      and only makes a difference when sum(ss) * sum(mumu) is nearly equal to
      sum(mus)**2. Overall divergence should be less than 1e-17.
}

