% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cgf_ratio}
\alias{cgf_ratio}
\title{get CGF-like ratio}
\usage{
cgf_ratio(lambdas, weights, mu)
}
\arguments{
\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}

\item{mu}{overall estimated mutation rate}
}
\value{
float value for
}
\description{
get CGF-like ratio
}

