% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{compute_pvalue}
\alias{compute_pvalue}
\title{compute de novo enrichment for a gene}
\usage{
compute_pvalue(de_novos, n_male, n_female, symbol, severity, mu_rate)
}
\arguments{
\item{de_novos}{table of de novos}

\item{n_male}{number of males}

\item{n_female}{number of females}

\item{symbol}{HGNC symbol for a gene}

\item{severity}{table of per base and per allele severity scores, for every gene}

\item{mu_rate}{table of mutation rates}
}
\value{
vector of p-value of de novo enrichment for a gene
}
\description{
compute de novo enrichment for a gene
}

