% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_approximation.R
\name{conditional_approximation}
\alias{conditional_approximation}
\title{conditional approximation}
\usage{
conditional_approximation(x, y, lambdas, weights)
}
\arguments{
\item{x}{estimated summed score}

\item{y}{observed summed score}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}
}
\value{
approximate value
}
\description{
conditional approximation
}

