% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_saddle_point.R
\name{double_saddle_point_approximation}
\alias{double_saddle_point_approximation}
\title{estimate gene-wise de novo enrichment by saddlepoint approximation}
\usage{
double_saddle_point_approximation(y, lambdas, weights)
}
\arguments{
\item{y}{this is the summed score for the observed de novos}

\item{lambdas}{vector of per base and allele mutation rates within a gene}

\item{weights}{vector of per base and allele weights, indicating the likely
severity of each change.}
}
\value{
p-value for gene
}
\description{
estimate gene-wise de novo enrichment by saddlepoint approximation
}

