% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment.R
\name{exclude_and_get_vector}
\alias{exclude_and_get_vector}
\title{exclude certain values from a table ,and convert to a vector}
\usage{
exclude_and_get_vector(data, exclude)
}
\arguments{
\item{data}{table of values}

\item{exclude}{matching table of booleans, where TRUE indicates values to exclude}
}
\value{
vector of values, with NA values removed
}
\description{
exclude certain values from a table ,and convert to a vector
}

