% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_rows_to_exclude}
\alias{get_rows_to_exclude}
\title{identify exclusion rows}
\usage{
get_rows_to_exclude(data)
}
\arguments{
\item{data}{dataframe of either mutation rates per site, or severity scores
per site.}
}
\value{
vector of booleans, where TRUE indicates a row to exclude
}
\description{
identify exclusion rows
}

