% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_approximation.R
\name{get_w}
\alias{get_w}
\title{w part function}
\usage{
get_w(s, s0, x, y, mu, lambdas, weights)
}
\arguments{
\item{s}{something}

\item{s0}{something}

\item{x}{tested summed score}

\item{y}{observed summed score}

\item{mu}{mutation rate}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}
}
\value{
summed value
}
\description{
w part function
}

