% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_approximation.R
\name{refine_w}
\alias{refine_w}
\title{refine w part until abs(w_part) is > 1e-4}
\usage{
refine_w(x, y, lambdas, weights, s0, increment = 0.01)
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}

\item{s0}{initial s value}

\item{increment}{value to increment y value per iteration}
}
\value{
list of s, mu and w_part values
}
\description{
refine w part until abs(w_part) is > 1e-4
}

