% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_approximation.R
\name{saddlepoint_p}
\alias{saddlepoint_p}
\title{calculate the p-value using the double saddle point}
\usage{
saddlepoint_p(values, s0, lambdas, weights)
}
\arguments{
\item{values}{list of 's', 'mu' and 'w_part' values}

\item{s0}{initial s value}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}
}
\value{
p-value
}
\description{
calculate the p-value using the double saddle point
}

