% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_s_u}
\alias{solve_s_u}
\title{solve for S and mu}
\usage{
solve_s_u(x, y, lambdas, weights, delta = 10, refine = 5, start = 0)
}
\arguments{
\item{x}{estimated summed score}

\item{y}{observed summed score}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}

\item{delta}{amount to update by during iterations}

\item{refine}{number of iteratative refinements to perform}

\item{start}{initial mu estimate}
}
\value{
list with 'mu' and 's' entries
}
\description{
solve for S and mu
}

