% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver}
\alias{solver}
\title{solve for initial and updated}
\usage{
solver(current, initial_sign, ratio, lambdas, weights, delta)
}
\arguments{
\item{current}{currently estimated mutation rate}

\item{initial_sign}{sign of initial estimate}

\item{ratio}{ratio of estimated and observed summed scores}

\item{lambdas}{vector of per base and allele mutation rates}

\item{weights}{per base and allele weights, indicating the likely severity of
each change}

\item{delta}{amount to update by during iterations}
}
\value{
list of inital and updated values
}
\description{
solve for initial and updated
}

