% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{tidy_table}
\alias{tidy_table}
\title{prepare dataframe for analysis}
\usage{
tidy_table(data, exclude)
}
\arguments{
\item{data}{dataframe of either mutation rates per site, or severity scores
per site.}

\item{exclude}{vector of booleans, where TRUE indicates a row to exclude}
}
\value{
dataframe with  columns renamed, and unecessary rows excluded
}
\description{
This changes some columns names to lower case, and excludes some predefined
rows. We want to exclude rows without severity scores. This function is then
applied for both the severity scores and mutation rates dataframes, so as to
keep the dataframes consistent.
}

